pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.poetry]
package-mode = false

[project]
name = "opendate"
version = "0.1.33"
description = "Python business datetimes"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "bissli", email = "bissli.xyz@protonmail.com" }]
requires-python = ">=3.9"
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "pandas-market-calendars",
  "pendulum",
  "wrapt",
  "typing-extensions",
]

[project.optional-dependencies]
test = [
  "asserts",
  "pdbpp",
  "pytest",
  "bump2version",
]

[project.urls]
Repository = "https://github.com/bissli/opendate"

[tool.maturin]
module-name = "opendate._opendate"
features = ["pyo3/extension-module"]
python-packages = ["opendate"]
python-source = "src"
manifest-path = "rust/Cargo.toml"
